+++ include/CPISync/Data/DataObject.h
@@ -121,7 +121,7 @@ public:
      * @return A char array version of the contents of this data object.
      *         The string could have null bytes and non-printable characters.
      */
-    const char *to_char_array(long &len) const;
+    const char *to_char_array(size_t &len) const;
 
     /**
      * @return A string of printable characters representing the object, essentially
